mosamatic-insights/pyproject.toml
[project]
name = "mosamatic-insights"
version = "0.1.0"
description = ""
authors = [
    {name = "Brecheisen", email = "r.brecheisen@maastrichtuniversity.nl"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pyside6-essentials>=6.9",
    "pydicom>=3.0.1",
    "numpy==2.4.1",
    "gdcm",
    "pylibjpeg>=2.0",
    "pylibjpeg-openjpeg>=2.0",
    "pillow>=10.0",
]

[tool.poetry]
packages = [{include = "mosamaticinsights", from = "src"}]

#[[tool.poetry.source]]
#name = "pytorch-cu121"
#url = "https://download.pytorch.org/whl/cu121"
#priority = "supplemental"

[tool.poetry.scripts]
mosamatic-insights = "mosamaticinsights.app:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pyinstaller = "^6.17.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
